requirements.txt
netcdf4
scipy
numpy
pyproj
geopandas
lxml
pyyaml
sentinelsat
